cranelift: Merge `abi` and `abi_impl` modules
