+++ cranelift/codegen/src/isa/x64/abi.rs
@@ -4,7 +4,7 @@ use crate::ir::{self, types, LibCall, MemFlags, Opcode, Signature, TrapCode, Typ
 use crate::ir::{types::*, ExternalName};
 use crate::isa;
 use crate::isa::{unwind::UnwindInst, x64::inst::*, x64::settings as x64_settings, CallConv};
-use crate::machinst::abi_impl::*;
+use crate::machinst::abi::*;
 use crate::machinst::*;
 use crate::settings;
 use crate::{CodegenError, CodegenResult};
