+++ cranelift/codegen/src/machinst/mod.rs
@@ -73,8 +73,6 @@ pub mod blockorder;
 pub use blockorder::*;
 pub mod abi;
 pub use abi::*;
-pub mod abi_impl;
-pub use abi_impl::*;
 pub mod buffer;
 pub use buffer::*;
 pub mod helpers;
